+++ src/pmap/__init__.py
@@ -1,3 +1,3 @@
 from .core import *
 
-__all__ = ['pmap', 'Future']
+__all__ = ['pmap', 'pvalmap', 'pkeymap', 'Deferred', 'Future']