+++ src/pmap/core.py
@@ -2,6 +2,8 @@ import multiprocessing
 import multiprocessing.pool
 import sys
 import threading
+from itertools import izip
+from toolz.compatibility import iterkeys, itervalues
 
 
 class Deferred(object):
@@ -85,6 +87,9 @@ def pmap(f, seq, threads=None, timeout=None):
 
     By default will spawn one thread per available core, but for I/O bound tasks it might be useful to define
     `threads` higher, to saturate throughput.
+
+    Use `timeout` argument to avoid blocking on an element of the sequence indefinitely. Will throw
+    `multiprocessing.TimeoutError` after the specified number of seconds.
     """
     assert (threads is None) or (type(threads) is int)
     assert (timeout is None) or (type(timeout) is int)
@@ -94,3 +99,31 @@ def pmap(f, seq, threads=None, timeout=None):
     fpool = [Future(f, pool=pool)(ele) for ele in seq]
     # Return a lazy generator that will block for results as necessary
     return (f.deref(timeout) for f in fpool)
+
+
+def pvalmap(f, d, factory=dict, threads=None, timeout=None):
+    """
+    Apply `f` to each value of dictionary `d` and return a new dictionary.
+
+    Blocks until entire dictionary can be realized.
+
+    Takes a `factory` to allow a return type that implements `collections.MutableMapping` other than `dict`.
+    """
+    rv = factory()
+    rv.update(
+        izip(iterkeys(d), pmap(f, itervalues(d), threads, timeout)))
+    return rv
+
+
+def pkeymap(f, d, factory=dict, threads=None, timeout=None):
+    """
+    Apply `f` to each key of dictionary `d` and return a new dictionary.
+
+    Blocks until entire dictionary can be realized.
+
+    Takes a `factory` to allow a return type that implements `collections.MutableMapping` other than `dict`.
+    """
+    rv = factory()
+    rv.update(
+        izip(pmap(f, iterkeys(d), threads, timeout), itervalues(d)))
+    return rv